requirements.txt
pydantic
python-dotenv
rapidfuzz
psycopg2
sqlalchemy
fastapi
groq